pyproject.toml
[project]
name = "lemonaid"
version = "0.6.0"
description = "Attention inbox for managing notifications from LLMs and other background tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.89.0",
]

[project.optional-dependencies]
# Requires macOS with Accessibility permissions granted
macos = [
    "pyobjc-framework-Quartz>=12.0",
    "pyobjc-framework-ApplicationServices>=12.0",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "ruff",
]

[project.scripts]
lemonaid = "lemonaid.cli:main"
lma = "lemonaid.cli:inbox_main"
lemonaid-tmux-window-status = "lemonaid.tmux.window_status:main"
lemonaid-tmux-session-name = "lemonaid.tmux.window_status:session_main"
lemonaid-claude-statusline = "lemonaid.claude.statusline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lemonaid"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "RUF012", # mutable class attrs - Textual uses this pattern
    "RUF022", # __all__ not sorted - we organize by category
]

[tool.ruff.lint.isort]
known-first-party = ["lemonaid"]
